requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
sentence-transformers
numpy
pymupdf
python-docx
pydantic